fix gradio and folder name
